Load the `PATH` environment variable in `~/.profile`
This is needed because GDM only loads this file after logging in, which means
any programs started from Gnome will only inherit these settings.

However, we'll still keep configuring this in `.bashrc` and `.zshrc` as well,
since not all shells load `.profile`.

This is all very entangled and the available documentation is not at all
readable or user-friendly, so it's not easy to get it right without a ton of
research.