Remove scripts for running Gulp, PHPCS, and PHPUnit
The scripts were meant to enable automatically running either the
locally-installed binary within a project, or fallback to a globally-installed
binary otherwise. However, these tools are nowadays almost exclusively executed
within Docker containers, so it makes little sense to use these scripts.

If one of them needs to be run on the host machine, it's easy enough to do so
manually.